+++ src/build_number.h
@@ -1,1 +1,1 @@
-#define BUILD 536
+#define BUILD 537